Code splitting and cleaning
